+++ apps/client/src/app/lazy-data/model/lazy-instance.ts
@@ -1,8 +1,25 @@
 export interface LazyInstance {
-  en:           null;
-  ja:           null;
-  de:           null;
-  fr:           null;
-  icon:         null;
+  en:           null | string;
+  ja:           null | string;
+  de:           null | string;
+  fr:           null | string;
+  icon:         Icon | null;
   contentText?: number[];
+}
+
+export enum Icon {
+  I061000061801PNG = "/i/061000/061801.png",
+  I061000061802PNG = "/i/061000/061802.png",
+  I061000061803PNG = "/i/061000/061803.png",
+  I061000061804PNG = "/i/061000/061804.png",
+  I061000061805PNG = "/i/061000/061805.png",
+  I061000061806PNG = "/i/061000/061806.png",
+  I061000061808PNG = "/i/061000/061808.png",
+  I061000061815PNG = "/i/061000/061815.png",
+  I061000061820PNG = "/i/061000/061820.png",
+  I061000061823PNG = "/i/061000/061823.png",
+  I061000061824PNG = "/i/061000/061824.png",
+  I061000061832PNG = "/i/061000/061832.png",
+  I061000061836PNG = "/i/061000/061836.png",
+  I061000061846PNG = "/i/061000/061846.png",
 }